(pC-1677) refacto Main and Header
