+++ src/components/layout/Header/mapStateToProps.js
@@ -1,0 +1,9 @@
+const mapStateToProps = state => {
+  const name = state.user && state.user.publicName
+  return {
+    name,
+    offerers: state.data.offerers,
+  }
+}
+
+export default mapStateToProps
