+++ src/components/layout/Main/mapStateToProps.js
@@ -1,0 +1,9 @@
+import { selectCurrentUser } from 'with-login'
+
+function mapStateToProps(state) {
+  return {
+    currentUser: selectCurrentUser(state),
+  }
+}
+
+export default mapStateToProps
